refactor:TableData, Order, and HeadCell types moved to type file
